% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanpath.R
\name{avg.group.dist}
\alias{avg.group.dist}
\title{Mean similarities among sets of scanpaths}
\usage{
avg.group.dist(d, groups)
}
\arguments{
\item{d}{a symmetric matrix of similarities as calculated by the
function \code{\link{scasim}} or a dist object (see
\code{\link{dist}}).}

\item{groups}{a vector that assigns each column to a group.}
}
\value{
a matrix containing the pair-wise average similarities for
all groups.
}
\description{
Calculates the mean similarities among sub-sets of scanpaths given
by a grouping variable.  For instance, when the groups are the
participants of the experiment, we get a matrix of the average
similarities between the scanpaths of the subjects.
}
\examples{
data(eyemovements)
d <- scasim(eyemovements, dur ~ x + y | trial, 512, 384, 60, 1/30)
s <- constant.vars(eyemovements, trial)$subject
avg.group.dist(d, s)
}
\seealso{
\code{\link{scasim}}, \code{\link{dist}}
}

