% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpath.R
\name{constant.vars}
\alias{constant.vars}
\title{Extract variables that are constant within the levels of a
grouping variable.}
\usage{
constant.vars(data, groups)
}
\arguments{
\item{data}{a data frame.}

\item{groups}{a grouping variable.  Either a vector with the same length as the data frame or simply a column in \code{data}.}
}
\value{
A data frame with one entry for each level of the grouping variable.
}
\description{
Given a data frame and a grouping variable, this function returns a
new data frame that contains only the variables that are constant
within the levels of the grouping variable.  The results data frame
contains only one line for each level of the grouping
variable.  This function can be used to extract trial-based
information from a data frame containing experimental data with
multiple measurements per trial.
}
\examples{
data(eyemovements)
head(eyemovements)
constant.vars(eyemovements, trial)
}
