% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanpath.R
\name{inverse.gnomonic}
\alias{inverse.gnomonic}
\title{Project points from a plane to a sphere.}
\usage{
inverse.gnomonic(x, y, center_x, center_y, distance, unit_size = 1)
}
\arguments{
\item{x}{the x-coordinates of the points.}

\item{y}{the y-coordinates of the points.}

\item{center_x}{is the x-coodrinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{center_y}{is the y-coodrinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{distance}{is the distance of the plane to the center of the
sphere.}

\item{unit_size}{is ratio of one unit of the plane coordinate
system and the unit in which the viewing distance was given.  So,
one unit in the coordinate system is \code{unit_size * unit of
distance}.}
}
\value{
A data frame containing lat and lon coordinates.
}
\description{
Projects from points given in plane coordinates to lat-lon, i.e.,
to the surface of a sphere (inverse gnomonic projection).  This can
be used to project screen coordinates to coordinates of the visual
field of somebody sitting in front of the screen.
}
\examples{
data(eyemovements)
x <- eyemovements$x
y <- eyemovements$y

latlon <- inverse.gnomonic(x, y, 268, 382, 8, 1/30)

# Before projection:
plot(x, y)
# After:
with(latlon, plot(lon, lat))
}
\references{
See
http://mathworld.wolfram.com/GnomonicProjection.html.
}

