% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scanpaths.R
\name{plot_alignment}
\alias{plot_alignment}
\title{This function plots the alignment of fixations in two scanpaths}
\usage{
plot_alignment(s, t, a, formula, offset_x = 0, offset_y = 0, nudge_x = 0,
  nudge_y = 0)
}
\arguments{
\item{s}{is a data frame containing the first scanpath.  Each line
represents one fixation.  The fixations have to be listed in
chronological order.  Required columns are: x- and y- coordinate
of fixations, and fixation durations.  See the example data set
provided with this package.}

\item{t}{is the second scanpath.  Same format as for s.}

\item{a}{is the alignment data frame as returned by
\code{\link{rscasim}}.}

\item{formula}{specifies which columns in the given data frame are
should be used.  The left-hand side specifies the column that
contains the fixation duration.  The right-hand side consist of
terms that specify x- and y-coordinate of the fixations (in that
order).  See examples below.}

\item{offset_x}{offset in the x dimension used for scanpath t.  Can
be used to reduce clutter.}

\item{offset_y}{like \code{offset_x} but in the y dimension.}

\item{nudge_x}{x offset used for fixation indices.}

\item{nudge_y}{x offset used for fixation indices.}
}
\value{
A ggplot object.  Can be modified before display, for
  example, by adding axis limits, labels, and titles, etc.  See
  examples below.
}
\description{
This function plots the alignment of fixations in two scanpaths
}
\examples{
data(eyemovements)
s <- subset(eyemovements, trial==1)
t <- subset(eyemovements, trial==9)
alignment <- rscasim(s, t, duration ~ x + y | trial,
                     512, 384, 60, 1/30)
plot_alignment(s, t, alignment, duration ~ x + y | trial, 10, 20)
}
