% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixsel.R
\name{replace.all}
\alias{replace.all}
\title{Replace all values in a vector by other values}
\usage{
replace.all(x, r, res.type = as.character, na.value = NA, ...)
}
\arguments{
\item{x}{a vector.}

\item{r}{a vector of replacement values.  One value for each
element in \code{dot(unique(x))}.}

\item{res.type}{a function for converting the result vector to a
desired data type.}

\item{na.value}{a replacement value for NAs.}

\item{...}{extra arguments that are passed to \code{res.type}.}
}
\value{
a version of \code{x} in which all original values are
substituted by the replacements specified in \code{r}.
}
\description{
Given a vector of values \code{x} and a set of replacement values
\code{r} (one for each unique value in \code{x}), this function
replaces each value in \code{x} by the corresponding new value in
\code{r}.
}
\note{
This function can be used to replace arbitrary region
identifiers in fixation data by single letters.  This format is
needed by the functions \code{\link{find.fixation}} and
\code{\link{match.scanpath}}.
}
\examples{
replace.all(1:10, letters[1:10])
replace.all(10:1, letters[1:10])

data(eyemovements)
words <- eyemovements$word

replace.all(words, letters[1:length(unique(words))])
replace.all(words, letters[1:length(unique(words))], paste, collapse="")
}
