% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpath.R
\name{rscasim}
\alias{rscasim}
\title{Calculate the similarity of two scanpaths}
\usage{
rscasim(s, t, formula, center_x, center_y, viewing_distance, unit_size,
  modulator = 0.83)
}
\arguments{
\item{s}{is a data frame containing the first scanpath.  Each line
represents one fixation.  The fixations have to be listed in
chronological order.  Required columns are: x- and y- coordinate
of fixations, and fixation durations.  See the example data set
provided with this package.}

\item{t}{is the second scanpath.  Same format as for s.}

\item{formula}{specifies which columns in the given data frame are
should be used.  The left-hand side specifies the column that
contains the fixation duration.  The right-hand side consist of
terms that specify x- and y-coordinate of the fixations (in that
order).  See examples below.}

\item{center_x}{is the x-coordinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{center_y}{is the y-coordinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{viewing_distance}{is the distance of the eyes from the screen
in some arbitrary unit.}

\item{unit_size}{the ratio of one unit of the coordinate system to
one of the units in which the viewing distance was
given.  Example: If the coordinates are pixels on a screen with
60 dpi and the unit of the distance is inches, \code{unit_size}
has to be set to 1/60.}

\item{modulator}{specifies how spatial distances between fixations
are assessed.  When set to 0, any spatial divergence of two
compared fixations is penalized independently equally
strong.  When set to 1, the scanpaths are compared only with
respect to their temporal patterns and spatial distances between
matching pairs of fixations are ignored.  The default value
mimics the exponential drop-off in human visual acuity as we move
away from the center from the fovea.}
}
\value{
A data frame showing how fixations in the two scanpaths
  were aligned.  Columns \code{s} and \code{t} contain the indices
  of the fixations in scanpaths s and t, and the column ~cost~ the
  cost of the edit operation.  The sum of the costs is the total
  dissimilarity of s and t.  If either \code{s} or \code{t} contain
  an NA, that means that a fixation in one scanpath didn't have a
  matching counterpart in the other scanpath.
}
\description{
Given two scanpaths, this function calculates their
dissimilarity.  This is a slow pure-R implementation of the scasim
measure written provided for educational purposes.  For analyses of
larger data sets function \code{\link[scanpath]{scasim}} should be
used.
}
\examples{
data(eyemovements)
s <- subset(eyemovements, trial==1)
t <- subset(eyemovements, trial==9)
rscasim(s, t, duration ~ x + y | trial,
        512, 384, 60, 1/30)
}
\references{
von der Malsburg, T. and Vasishth, S. (2011). What is the scanpath
signature of syntactic reanalysis? Journal of Memory and Language,
65(2):109-127.

von der Malsburg, T., Vasishth, S., and Kliegl,
R. (2012). Scanpaths in reading are informative about sentence
processing. In Michael Carl, P. B. and Choudhary, K. K., editors,
Proceedings of the First Workshop on Eye-tracking and Natural
Language Processing, pages 37-53, Mumbai, India. The COLING 2012
organizing committee.
}
\seealso{
\code{\link[scanpath]{scasim}}, \code{\link[scanpath]{plot_alignment}}
}
\author{
Titus von der Malsburg \email{malsburg@posteo.de}
}
\keyword{cluster}
\keyword{eye}
\keyword{movements}
\keyword{scanpaths}
