\name{scanpath-package}

\alias{scanpath-package}

\alias{scanpath}

\docType{package}

\title{Tools for analyzing eye movements}

\description{Provides tools for (pre-)processing and analyzing eye movement
data.  Currently, this package contains the similarity measure introduced by
von der Malsburg \& Vasishth (2007).  It serves as a basis of the analysis of
temporal and spatial patterns in sequences of fixations using clustering
methods and multi-dimensional scaling.}

\details{
\tabular{ll}{
Package: \tab scanpath\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-11-12\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
See \code{\link{scasim}} and \code{\link{inverse.gnomonic}} for further details.
}

\author{Titus von der Malsburg <malsburg@uni-potsdam.de>}

\references{Titus von der Malsburg and Shravan Vasishth (2007) A
time-sensitive similarity measure for scanpaths, in Proceedings of the European
Conference on Eye Movements, Potsdam, Germany}

\keyword{ package }
\keyword{ cluster }
\keyword{ ts }
\keyword{ htest }
\seealso{ See also \code{\link{scasim}} and \code{\link{inverse.gnomonic}}.}
