% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpath.R
\docType{package}
\name{scanpath}
\alias{scanpath}
\alias{scanpath-package}
\title{Tools for analyzing scanpaths}
\description{
This package provides tools for analyzing spatio-temporal patterns
in eye movements, a.k.a. scanpaths.
}
\references{
von der Malsburg, T. and Vasishth, S. (2011). What is the scanpath
signature of syntactic reanalysis? Journal of Memory and Language,
65(2):109-127.
}
\seealso{
\code{\link{scasim}}, \code{\link{avg.group.dist}}, \code{\link{constant.vars}}, \code{\link{which.centroid}}
}
\author{
Titus von der Malsburg \email{malsburg@posteo.de}
}
\keyword{cluster}
\keyword{eye}
\keyword{movements}
\keyword{package}
\keyword{scanpaths}
