% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanpath.R
\name{which.centroid}
\alias{which.centroid}
\title{Find item with minimal average distance to other items}
\usage{
which.centroid(d, select = NULL)
}
\arguments{
\item{d}{a symmetric matrix of similarities or a \code{\link{dist}}
object.}

\item{select}{a vector of names of items that should be included in
the analysis.  These names should correspond to the row and column
names of \code{d}.  Items not listed are ignored completely.  The
default is to take all items into account.}
}
\value{
the name of the item that has the minimal average distance
to the other items.
}
\description{
Given a matrix of similarities, this function identifies the item
that has the smallest average distance to the other items.
}
\details{
Although the function is called which.centroid, the
selected item is not necessarily the centroid in the strict
sense.  However, it is the item which is closer to the centroid
that any other item in the set.
}
\examples{
data(eyemovements)
d <- scasim(eyemovements, dur ~ x + y | trial, 512, 384, 60, 1/30)
which.centroid(d)
which.centroid(d, c("1:1", "1:2", "1:3"))
}

