% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixsel.R
\name{find.fixation}
\alias{find.fixation}
\title{Mark the beginning of specified pattern in a letter sequence}
\usage{
find.fixation(l, groups, expr, nth = NA, subpattern = 0)
}
\arguments{
\item{l}{a vector of single characters.}

\item{groups}{a grouping variable.  The search for matching
patterns in performed in each group separately.  Matches crossing
group boundaries are not considered.}

\item{expr}{a regular expression describing the pattern.}

\item{nth}{specifies which match should be returned if there were
several within a group.  The default is to return all matches.}

\item{subpattern}{the subpattern of interest.  If zero, the
beginning of the full pattern will be marked.  If n>0, the
beginning of the n-th subpattern will be marked.  See examples.}
}
\value{
A vector giving the indices in l where the matches
begin.  If subpattern > 0, then the indices will point to the
beginning of the subpatterns.
}
\description{
This function uses regular expressions to identify elements in a
vector of letters where a specific pattern begins.  Among other
things, this can be used to identify certain eye movement patterns
in fixation data.
}
\examples{
data(eyemovements)
words <- eyemovements$word
l <- replace.all(words, letters[1:length(unique(words))])
trial <- eyemovements$trial
# Find fixations on word 6 (which is represented by letter "f"):
idx <- find.fixation(l, trial, "f")
words[idx]
# Find fixations on word 6 but only the second match within
# a group (i.e., trial):
idx <- find.fixation(l, trial, "f", nth=2)
words[idx]
# Find fixations on word 6 that are followed by fixations on word 7 ("g"):
idx <- find.fixation(l, trial, "fg")
words[idx]
# Find those fixations on word 6 that followed fixations on word 5 ("e"):
idx <- find.fixation(l, trial, "e(f)", subpattern=1)
words[idx]
# Find the beginning of a second sweep over the sentence:
idx <- find.fixation(l, trial, "a[b-h]+(a+[b-g]+)", subpattern=1)
words[idx]
}
