% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixsel.R
\name{find.fixation}
\alias{find.fixation}
\title{Find specific fixations}
\usage{
find.fixation(l, groups, expr, nth = NA, subpattern = 0)
}
\arguments{
\item{l}{a vector of single letters or single-digit integers each
representing a fixation on a region of interest.}

\item{groups}{a grouping variable indicating to which scanpath each
fixation belongs.  The search for matching patterns in performed
within each group separately and matches crossing group boundaries
are not considered.}

\item{expr}{a regular expression describing the scanpath pattern of
interest.}

\item{nth}{specifies which match should be returned when there are
multiple within a group.  The default is to return all matches.}

\item{subpattern}{the subpattern of interest.  If zero, the
beginnings of the full patterns will be returned.  If n>0, the
beginnings of the n-th subpatterns will be returned.}
}
\value{
A vector giving the indices of the matches in l.
}
\description{
Finds fixations embedded in a context that is specified using a
regular expression.
}
\examples{
data(eyemovements)
words <- eyemovements$word
trial <- eyemovements$trial

# Find fixations on word 6 (which is represented by letter "f"):
idx <- find.fixation(words, trial, "6")
eyemovements[idx,]

# Same, but use letters as region identifiers, which is necessary
# when there are more than 10 regions.  (Code assumes that regions
# are numbered in ascending order starting from 1.)
w <- letters[match(words, 1:26)]
idx <- find.fixation(w, trial, "f")
eyemovements[idx,]

# Find fixations on word 6 but only the second match within
# a group (i.e., trial):
idx <- find.fixation(words, trial, "6", nth=2)
eyemovements[idx,]

# Find fixations on word 6 that are followed by fixations on word 7:
idx <- find.fixation(words, trial, "67")
eyemovements[idx,]

# Find those fixations on word 6 that are preceded by fixations on
# word 5:
idx <- find.fixation(words, trial, "5(6)", subpattern=1)
eyemovements[idx,]

# Find the beginning of a second sweep over the sentence:
idx <- find.fixation(words, trial, "[123]+[456]+[78]+([123]+[456]+[78]+)", subpattern=1)
eyemovements[idx,]
}
