% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixsel.R
\name{match.scanpath}
\alias{match.scanpath}
\title{Match scanpath patterns}
\usage{
match.scanpath(l, groups, expr, subpattern = 0)
}
\arguments{
\item{l}{a vector of single characters.}

\item{groups}{a grouping variable.  The search for matching
patterns in performed in each group separately.  Matches crossing
group boundaries are not considered.  This vector will typically
contain a trial id.}

\item{expr}{a regular expression describing the target pattern.}

\item{subpattern}{the subpattern of interest.  If zero, the
complete match will be marked.  If n>0, only the n-th subpattern
will be marked.  See examples.}
}
\value{
A vector giving the indices of the matching
  fixations.  Only the first match within each group will be
  marked.
}
\description{
This function matches contiguous sequences of fixations that follow
a specified pattern.  This target pattern is specified using
regular expressions.
}
\examples{
data(eyemovements)
words <- eyemovements$word
l <- replace.all(words, letters[1:length(unique(words))])
trial <- eyemovements$trial
# Find fixations on word 6 (which is represented by letter "f"):
idx <- match.scanpath(l, trial, "f")
}

