% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpath.R
\name{scasim}
\alias{scasim}
\title{Calculate pair-wise similarities among a set of scanpaths}
\usage{
scasim(data, formula, center_x, center_y, viewing_distance, unit_size,
  modulator = 0.83, data2 = NULL, formula2 = formula, normalize = FALSE)
}
\arguments{
\item{data}{is a data frame containing the eye movement data.  Each
line represents one fixation of the eyes.  The fixations of a trial
have to be listed en bloc and in chronological order.  Required
columns are: trial ID (unique in the whole data set), x- and y-
coordinate of fixations, and the fixation durations.  See the
example data set provided with this package.}

\item{formula}{specifies which columns in the given data frame are
relevant.  The left-hand side specifies the column that contains
the fixation duration The right-hand side consist of terms that
specify x- and y-coordinate of the fixations (in that order)
conditioned on trial IDs.  See examples below.}

\item{center_x}{is the x-coordinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{center_y}{is the y-coordinate of the point that is targeted
when the eye looks straight ahead (usually the center of the
screen).}

\item{viewing_distance}{is the distance of the eyes to the
screen.}

\item{unit_size}{is ratio of one unit of the coordinate system and
the unit in which the viewing distance was given.  So, one unit in
the coordinate system is \code{unit_size * unit of
distance}.  Example: If the coordinates are pixels on a screen with
60 dpi and the unit of the distance is inches, \code{unit_size} has
to be set to 1/60.}

\item{modulator}{specifies how spatial distances between fixations
are assessed.  When set to 0, any spatial divergence of two
compared scanpaths is penalized independently of its degree.  When
set to 1, the scanpaths are compared only with respect to their
temporal patterns.  The default value approximates the sensitivity
to spatial distance found in the human visual system.}

\item{data2}{optionally provides a second set of scanpaths.  In
this case \code{scasim} computes the distances from each scanpath
in the first set to each scanpath in the second set.  The resulting
matrix of distances is therefore not necessarily square.}

\item{formula2}{specifies the relevant columns in \code{data2} in
case they are named differently in \code{data2} than in
\code{data}.  \code{formula2} defaults to \code{formula}.}

\item{normalize}{specifies how each calculated similarity value for
scanpaths s and t should be normalized.  There are three possible
modes:

\describe{
  \item{\code{fixations}}{normalize by dividing by the total number
  of fixations in s and t,}
  \item{\code{durations}}{divide by the total duration of s and t,}
  \item{\code{FALSE}}{don't normalize at all.}
}

The choice of normalization mode can have a strong effect when
fitting maps of scanpaths using \code{\link[MASS]{isoMDS}} or
\code{\link{cmdscale}}.  Not normalizing at all can yield maps that
are difficult to interpret when scanpaths differ markedly in their
duration or number of fixations.}
}
\value{
A matrix containing the pair-wise dissimilarities.  Columns
and rows correspond to trials.  Their order is the same as the
order of trials in \code{data}.
}
\description{
Given a set of scanpaths, this function calculates the similarities
between all pairs of scanpaths in the set.
}
\details{
When using \code{scasim}, the dissimilarity of two
scanpaths is assessed based on the distances of their fixation
targets.  If the distances are really small, the dissimilarity will
mostly be a function of just the differences of the fixation
durations.  The evaluation of distance accounts for the particular
way the human visual system magnifies the center of the visual
field and the drop in acuity at its periphery.
}
\examples{
data(eyemovements)

# Calculating dissimilarities: when looking straight ahead the gaze
# targets the point with the coordinates (512,384), viewing
# distance (eye to screen) is 60 cm, 1 unit in the data (pixel) is
# 1/30 cm.

dissimilarities <- scasim(eyemovements, duration ~ x + y | trial,
                          512, 384, 60, 1/30)

# Using cmdscale for fitting a map:

map <- cmdscale(dissimilarities)
plot(map, cex=4)
text(map, labels=rownames(map))
}
\references{
von der Malsburg, T. and Vasishth, S. (2011). What is the scanpath
signature of syntactic reanalysis? Journal of Memory and Language,
65(2):109-127.

von der Malsburg, T., Vasishth, S., and Kliegl,
R. (2012). Scanpaths in reading are informative about sentence
processing. In Michael Carl, P. B. and Choudhary, K. K., editors,
Proceedings of the First Workshop on Eye-tracking and Natural
Language Processing, pages 37-53, Mumbai, India. The COLING 2012
organizing committee.
}
\seealso{
\code{\link[MASS]{isoMDS}} can be applied to the output of
\code{scasim} in order to fitting maps of scanpaths.
}
\author{
Titus von der Malsburg \email{malsburg@posteo.de}
}
\keyword{cluster}
\keyword{eye}
\keyword{movements}
\keyword{scanpaths}
